feat: redesign classifier interface (2)
