+++ src/safeds/ml/classical/classification/_logistic_classifier.py
@@ -1,0 +1,54 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Self
+
+from safeds._utils import _get_random_seed, _structural_hash
+
+from ._classifier import Classifier
+
+if TYPE_CHECKING:
+    from sklearn.base import ClassifierMixin
+    from sklearn.linear_model import LogisticRegression as sk_LogisticRegression
+
+    from safeds.data.labeled.containers import TabularDataset
+    from safeds.data.tabular.containers import Table
+
+
+class LogisticClassifier(Classifier):
+    """Regularized logistic regression for classification."""
+
+    # ------------------------------------------------------------------------------------------------------------------
+    # Dunder methods
+    # ------------------------------------------------------------------------------------------------------------------
+
+    def __init__(self) -> None:
+        super().__init__()
+
+        # Internal state
+        self._wrapped_classifier: sk_LogisticRegression | None = None
+
+    def __hash__(self) -> int:
+        return _structural_hash(
+            super().__hash__(),
+        )
+
+    # ------------------------------------------------------------------------------------------------------------------
+    # Template methods
+    # ------------------------------------------------------------------------------------------------------------------
+
+    def _check_additional_fit_preconditions(self, training_set: TabularDataset):
+        pass
+
+    def _check_additional_predict_preconditions(self, dataset: Table | TabularDataset):
+        pass
+
+    def _clone(self) -> Self:
+        return LogisticClassifier()
+
+    def _get_sklearn_model(self) -> ClassifierMixin:
+        from sklearn.linear_model import LogisticRegression as SklearnLogisticRegression
+
+        return SklearnLogisticRegression(
+            random_state=_get_random_seed(),
+            n_jobs=-1,
+        )
