--- src/safeds/ml/classical/classification/_logistic_regression.py
@@ -1,113 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-from safeds._utils import _structural_hash
-from safeds.ml.classical._util_sklearn import fit, predict
-
-from ._classifier import Classifier
-
-if TYPE_CHECKING:
-    from sklearn.base import ClassifierMixin
-    from sklearn.linear_model import LogisticRegression as sk_LogisticRegression
-
-    from safeds.data.labeled.containers import TabularDataset
-    from safeds.data.tabular.containers import Table
-
-
-class LogisticRegressionClassifier(Classifier):
-    """Regularized logistic regression."""
-
-    def __hash__(self) -> int:
-        return _structural_hash(
-            super().__hash__(),
-        )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-        # Internal state
-        self._wrapped_classifier: sk_LogisticRegression | None = None
-
-    def fit(self, training_set: TabularDataset) -> LogisticRegressionClassifier:
-        """
-        Create a copy of this classifier and fit it with the given training data.
-
-        This classifier is not modified.
-
-        Parameters
-        ----------
-        training_set:
-            The training data containing the feature and target vectors.
-
-        Returns
-        -------
-        fitted_classifier:
-            The fitted classifier.
-
-        Raises
-        ------
-        LearningError
-            If the training data contains invalid values or if the training failed.
-        TypeError
-            If a table is passed instead of a tabular dataset.
-        NonNumericColumnError
-            If the training data contains non-numerical values.
-        MissingValuesColumnError
-            If the training data contains missing values.
-        DatasetMissesDataError
-            If the training data contains no rows.
-        """
-        wrapped_classifier = self._get_sklearn_model()
-        fit(wrapped_classifier, training_set)
-
-        result = LogisticRegressionClassifier()
-        result._wrapped_classifier = wrapped_classifier
-        result._feature_schema = training_set.features.column_names
-        result._target_name = training_set.target.name
-
-        return result
-
-    def predict(self, dataset: Table | TabularDataset) -> TabularDataset:
-        """
-        Predict a target vector using a dataset containing feature vectors. The model has to be trained first.
-
-        Parameters
-        ----------
-        dataset:
-            The dataset containing the feature vectors.
-
-        Returns
-        -------
-        table:
-            A dataset containing the given feature vectors and the predicted target vector.
-
-        Raises
-        ------
-        ModelNotFittedError
-            If the model has not been fitted yet.
-        DatasetMissesFeaturesError
-            If the dataset misses feature columns.
-        PredictionError
-            If predicting with the given dataset failed.
-        NonNumericColumnError
-            If the dataset contains non-numerical values.
-        MissingValuesColumnError
-            If the dataset contains missing values.
-        DatasetMissesDataError
-            If the dataset contains no rows.
-        """
-        return predict(self._wrapped_classifier, dataset, self._feature_schema, self._target_name)
-
-    def _get_sklearn_model(self) -> ClassifierMixin:
-        """
-        Return a new wrapped Classifier from sklearn.
-
-        Returns
-        -------
-        wrapped_classifier:
-            The sklearn Classifier.
-        """
-        from sklearn.linear_model import LogisticRegression as sk_LogisticRegression
-
-        return sk_LogisticRegression(n_jobs=-1)
